Data.Wheat.Codecs

Description

A collection of commonly-used codecs.

ByteStrings

byteString :: Codec ByteString

Encode a ByteString (the identity codec)

lazyByteString :: Codec ByteString

Encode a lazy ByteString (the identity codec)

constant :: ByteString -> Codec' ByteString e

Encode a constant ByteString.

Decoding will fail if the supplied ByteString is not a prefix of the actual one. Encoding will ignore its argument.

lazyConstant :: ByteString -> Codec' ByteString e

Encode a constant lazy ByteString.

Decoding will fail if the actual ByteString differs from the supplied one. Encoding will ignore its argument.

firstN :: Int -> Codec ByteString

Encode the first N bytes of a ByteString.

Decoding will fail if the actual ByteString is not at least as long as the given length.

lazyFirstN :: Int -> Codec ByteString

Encode the first N bytes of a lazy ByteString.

Decoding will fail if the actual ByteString is not at least as long as the given length.

codewords :: [(ByteString, ByteString)] -> Codec ByteString

Encode a ByteString using a collection of codewords, falling back to the identity function when a codeword is missing. Both encoding and decoding attempt to match codewords in order, allowing for non-prefix-free codes.

Lists

elementwise :: Codec' d e -> Codec' [d] [e]

Lift a codec over a type to over a list of that type.

Encoded elements are concatenated. Decoding never fails, as an empty list is acceptable.

Numbers

asciiDigits :: Codec Int

Encode an Int as ASCII digits.

Decoding will consume as many bytes from the start of the input as represent ASCII digits (with an optional single leading minus sign), failing only if no bytes do.