Data.Wheat.Combinators

Description

A collection of commonly-used combinators to build up larger codecs.

Documentation

(<:>) :: Monoid b => GCodec i b e d1 -> GCodec i b e d2 -> GCodec i b e (d1, d2)

Sequential composition of codecs.

When encoding, the input value is encoded with both codecs and the results are concatenated. When decoding, the remaining bytestring from the first codec is used as input to the second.

(<<:>>) :: (Monoid b, Monoid d) => GCodec i b e d -> GCodec i b e d -> GCodec i b e d

Sequential composition of codecs, combining the results of decoding monoidally.

(<:>>) :: Monoid b => GCodec i b e d1 -> GCodec i b e d2 -> GCodec i b e d2

Right-biased sequential composition of codecs.

Encoding behaviour is the same as <:>, decoding throws away the result of the first codec.

(<<:>) :: Monoid b => GCodec i b e d1 -> GCodec i b e d2 -> GCodec i b e d1

Left-biased sequential composition of codecs.

Encoding behaviour is the same as <:>, decoding throws away the result of the second codec.

(<++>) :: Monoid b => GCodec i b e1 d1 -> GCodec i b e2 d2 -> GCodec i b (e1, e2) (d1, d2)

Combine two codecs into a codec for tuples of those types, where the encoded forms of each element are concatenated.

elementwise :: Monoid b => GCodec i b e d -> GCodec i b [e] [d]

Lift a codec over a type to over a list of that type.

Encoded elements are concatenated. Decoding never fails, as an empty list is acceptable.

header :: Monoid b => (e -> h) -> GCodec i b h h -> (h -> GCodec i b e d) -> GCodec i b e d

Encode a value as a combination of header and encoded value. The codec used for encoding/decoding the value itself receives the (encoded/decoded) header as a parameter.

separate :: Monoid b => (e -> (e1, e2)) -> ((d1, d2) -> d) -> GCodec i b e1 d1 -> GCodec i b e2 d2 -> GCodec i b e d

Apply a divide-and-conquer approach: given a function to split up the encoding into two smaller components, and a function to combine the smaller components after decoding, construct a codec for the more complex type.

wrap :: (d' -> Maybe d) -> (e -> Maybe e') -> GCodec i b e' d' -> GCodec i b e d

Given functions to convert the types, wrap one codec inside another.

Encoding and decoding will fail if the function does, even if the inner codec succeeds.