Data.Wheat.Types

Description

Wheat is based around bidirectional codecs, although they are not constrained to deocde to the same type they decode (see Codec'). For the common case, these two type parameters will be the same.

Decoders are Monads, this means we can construct a decoder for a larger type out of decoders for smaller types along with a function to combine the parts. Encoders are Divisibles, this means we can construct an encoder for a larger type out of encoders for smaller types along with a function to split the whole.

Codecs

type Codec a = Codec' a a

A codec where the type we encode is the same as the type we decode to. Most of the primitive codec functions will have this type.

type Codec' e d = GCodec ByteString Builder e d

A more general codec where we are contravariant in the type we encode but covariant in the type we decode.

The Semigroup and Monoid instances represent left-biased choice. If the left codec fails, the right will be tried.

data GCodec i b e d

Constructors

Codec

Fields

encoderOf :: GEncoder b e

decoderOf :: GDecoder i d

Decoders

type Decoder a = GDecoder ByteString a

A decoder is a function which, given some input bytestring, will attempt to decode it according to the construction of the decoder and return any remaining input.

type GDecoder i a = StateT i Both a

toDecoder :: (i -> Maybe (a, i)) -> GDecoder i a

Construct a decoder.

toDecoder' :: (i -> Both (a, i)) -> GDecoder i a

Variant of toDecoder.

runDecoder :: GDecoder i a -> i -> Maybe (a, i)

Run a decoder.

runDecoder' :: GDecoder i a -> i -> Both (a, i)

Variant of runDecoder

decode :: GCodec i b e d -> i -> Maybe d

Run a codec's decoder.

Encoders

type Encoder a = GEncoder Builder a

An encoder is a function which, given some input value, will attempt to encode it according to the construction of the encoder, producing a ByteString builder.

type GEncoder b a = Op (Both b) a

toEncoder :: (a -> Maybe b) -> GEncoder b a

Construct an encoder.

toEncoder' :: (a -> Both b) -> GEncoder b a

Variant of toEncoder.

runEncoder :: GEncoder b a -> a -> Maybe b

Run an encoder.

runEncoder' :: GEncoder b a -> a -> Both b

Variant of runEncoder.

encode :: GCodec i b e d -> e -> Maybe b

Run a codec's encoder.