Data.Wheat

Description

A collection of commonly-used basic codecs and combinators for constructing larger codecs.

Re-exported types

module Data.Wheat.Types

ByteStrings

byteString :: Codec ByteString

Encode a ByteString (the identity codec)

lazyByteString :: Codec ByteString

Encode a lazy ByteString (the identity codec)

constant :: ByteString -> Codec' ByteString e

Encode a constant ByteString.

Decoding will fail if the supplied ByteString is not a prefix of the actual one. Encoding will ignore its argument.

lazyConstant :: ByteString -> Codec' ByteString e

Encode a constant lazy ByteString.

Decoding will fail if the actual ByteString differs from the supplied one. Encoding will ignore its argument.

firstN :: Int -> Codec ByteString

Encode the first N bytes of a ByteString.

Decoding will fail if the actual ByteString is not at least as long as the given length.

lazyFirstN :: Int -> Codec ByteString

Encode the first N bytes of a lazy ByteString.

Decoding will fail if the actual ByteString is not at least as long as the given length.

Lists

elementwise :: Codec' d e -> Codec' [d] [e]

Lift a codec over a type to over a list of that type.

Encoded elements are concatenated. Decoding never fails, as an empty list is acceptable.

Numbers

asciiDigits :: Codec Int

Encode an Int as ASCII digits.

Decoding will consume as many bytes from the start of the input as represent ASCII digits, failing only if no bytes do.

Combinators

(<:>) :: Codec' d1 e -> Codec' d2 e -> Codec' (d1, d2) e

Sequential composition of codecs.

When encoding, the input value is encoded with both codecs and the results are concatenated. When decoding, the remaining bytestring from the first codec is used as input to the second.

(<<:>>) :: Monoid d => Codec' d e -> Codec' d e -> Codec' d e

Sequential composition of codecs, combining the results of decoding monoidally.

(<:>>) :: Codec' d1 e -> Codec' d2 e -> Codec' d2 e

Right-biased sequential composition of codecs.

Encoding behaviour is the same as <:>, decoding throws away the result of the first codec.

(<<:>) :: Codec' d1 e -> Codec' d2 e -> Codec' d1 e

Left-biased sequential composition of codecs.

Encoding behaviour is the same as <:>, decoding throws away the result of the second codec.

(<++>) :: Codec' d1 e1 -> Codec' d2 e2 -> Codec' (d1, d2) (e1, e2)

Combine two codecs into a codec for tuples of those types, where the encoded forms of each element are concatenated.

dispatch :: Decoder d -> (e -> Builder) -> Codec' d e

Choose an encoder based on the actual value being encoded, with a decoder.

Be careful that this codec actually does express a reversible encoding when you have d == e!

header :: (e -> h) -> Codec h -> (h -> Codec' d e) -> Codec' d e

Encode a value as a combination of header and encoded value. The codec used for encoding/decoding the value itself receives the (encoded/decoded) header as a parameter.

separate :: (e -> (e1, e2)) -> ((d1, d2) -> d) -> Codec' d1 e1 -> Codec' d2 e2 -> Codec' d e

Apply a divide-and-conquer approach: given a function to split up the encoding into two smaller components, and a function to combine the smaller components after decoding, construct a codec for the more complex type.