Data.Wheat.Codecs

Description

A collection of commonly-used codecs.

ByteStrings

byteString :: Codec ByteString

Encode a ByteString (the identity codec)

lazyByteString :: Codec ByteString

Encode a lazy ByteString (the identity codec)

constant :: ByteString -> Codec' e ByteString

Encode a constant ByteString.

Decoding will fail if the supplied ByteString is not a prefix of the actual one. Encoding will ignore its argument.

lazyConstant :: ByteString -> Codec' e ByteString

Encode a constant lazy ByteString.

Decoding will fail if the actual ByteString differs from the supplied one. Encoding will ignore its argument.

firstN :: Int -> Codec ByteString

Encode the first N bytes of a ByteString.

Decoding will fail if the actual ByteString is not at least as long as the given length.

lazyFirstN :: Int -> Codec ByteString

Encode the first N bytes of a lazy ByteString.

Decoding will fail if the actual ByteString is not at least as long as the given length.

codewords :: [(ByteString, ByteString)] -> Codec ByteString

Encode a ByteString using a collection of codewords, falling back to the identity function when a codeword is missing. Both encoding and decoding attempt to match codewords in order, allowing for non-prefix-free codes.

Numbers

asciiDigits :: Codec Int

Encode an Int as ASCII digits.

Decoding will consume as many bytes from the start of the input as represent ASCII digits (with an optional single leading minus sign), failing only if no bytes do.

Miscellaneous

delimited :: Maybe ByteString -> Maybe ByteString -> Codec' e d -> Codec' e d

Surround a codec with delimiters.

If an ending delimiter is provided, and multiple instances of the ending delimiter crop up in the actual ByteString when decoding, the wrapped codec will be tried for all possible endings until decoding succeeds. If decoding succeeds but does not consume all the bytes between the starting and ending tokens, the overall decoding will fail.