Data.Wheat.Combinators

Description

A collection of commonly-used combinators to build up larger codecs.

Combinators

(<:>) :: Codec' d1 e -> Codec' d2 e -> Codec' (d1, d2) e

Sequential composition of codecs.

When encoding, the input value is encoded with both codecs and the results are concatenated. When decoding, the remaining bytestring from the first codec is used as input to the second.

(<<:>>) :: Monoid d => Codec' d e -> Codec' d e -> Codec' d e

Sequential composition of codecs, combining the results of decoding monoidally.

(<:>>) :: Codec' d1 e -> Codec' d2 e -> Codec' d2 e

Right-biased sequential composition of codecs.

Encoding behaviour is the same as <:>, decoding throws away the result of the first codec.

(<<:>) :: Codec' d1 e -> Codec' d2 e -> Codec' d1 e

Left-biased sequential composition of codecs.

Encoding behaviour is the same as <:>, decoding throws away the result of the second codec.

(<++>) :: Codec' d1 e1 -> Codec' d2 e2 -> Codec' (d1, d2) (e1, e2)

Combine two codecs into a codec for tuples of those types, where the encoded forms of each element are concatenated.

delimited :: Maybe ByteString -> Maybe ByteString -> Codec' d e -> Codec' d e

Surround a codec with delimiters.

If an ending delimiter is provided, and multiple instances of the ending delimiter crop up in the actual ByteString when decoding, the wrapped codec will be tried for all possible endings until decoding succeeds. If decoding succeeds but does not consume all the bytes between the starting and ending tokens, the overall decoding will fail.

dispatch :: Decoder d -> (e -> Builder) -> Codec' d e

Choose an encoder based on the actual value being encoded, with a decoder.

Be careful that this codec actually does express a reversible encoding when you have d == e!

header :: (e -> h) -> Codec h -> (h -> Codec' d e) -> Codec' d e

Encode a value as a combination of header and encoded value. The codec used for encoding/decoding the value itself receives the (encoded/decoded) header as a parameter.

separate :: (e -> (e1, e2)) -> ((d1, d2) -> d) -> Codec' d1 e1 -> Codec' d2 e2 -> Codec' d e

Apply a divide-and-conquer approach: given a function to split up the encoding into two smaller components, and a function to combine the smaller components after decoding, construct a codec for the more complex type.