Data.Wheat.Types

Description

Wheat is based around bidirectional codecs, although they are not constrained to deocde to the same type they decode (see Codec'). For the common case, these two type parameters will be the same.

Decoders are Monads, this means we can construct a decoder for a larger type out of decoders for smaller types along with a function to combine the parts. Encoders are Divisibles, this means we can construct an encoder for a larger type out of encoders for smaller types along with a function to split the whole.

Codecs

type Codec a = Codec' a a

A codec where the type we encode is the same as the type we decode to. Most of the primitive codec functions will have this type.

data Codec' d e where

A more general codec where we are covariant in the type we encode but contravariant in the type we decode.

There are two constructors,

Plain: The simplest type of codec is just a pair of a Decoder and Encoder.

Wrap: Sometimes it's useful to operate in terms of one type, but do transformation to another under the hood (see the bencoding example in the README), as this means we can have a collection of codecs which are conceptually the same type, but have very different implementations.

Constructors

Plain :: Decoder d -> Encoder e -> Codec' d e

Wrap :: (d' -> Maybe d) -> (e -> Maybe e') -> Codec' d' e' -> Codec' d e

Decoders

newtype Decoder a

A decoder is a function which, given some input bytestring, will attempt to decode it according to the construction of the decoder and return any remaining input.

Constructors

Decoder

Fields

runDecoder :: ByteString -> Maybe (a, ByteString)

decode :: ByteString -> Codec' d e -> Maybe d

Run a codec's decoder.

decoderOf :: Codec' d e -> Decoder d

Get the decoder of a codec.

Encoders

newtype Encoder a

An encoder is a function which, given some input value, will attempt to encode it according to the construction of the encoder, producing a ByteString builder.

Constructors

Encoder

Fields

runEncoder :: a -> Maybe Builder

encode :: e -> Codec' d e -> Maybe Builder

Run a codec's encoder.

encoderOf :: Codec' d e -> Encoder e

Get the encoder of a codec.