Data.Wheat.Types

Description

Wheat is based around bidirectional codecs, although they are not constrained to deocde to the same type they decode (see Codec'). For the common case, these two type parameters will be the same.

Decoders are Monads, this means we can construct a decoder for a larger type out of decoders for smaller types along with a function to combine the parts. Encoders are Divisibles, this means we can construct an encoder for a larger type out of encoders for smaller types along with a function to split the whole.

Codecs

type Codec a = Codec' a a

A codec where the type we encode is the same as the type we decode to. Most of the primitive codec functions will have this type.

type Codec' d e = GCodec ByteString Builder d e

A more general codec where we are covariant in the type we encode but contravariant in the type we decode.

There are two constructors,

Plain: The simplest type of codec is just a pair of a Decoder and Encoder.

Wrap: Sometimes it's useful to operate in terms of one type, but do transformation to another under the hood (see the bencoding example in the README), as this means we can have a collection of codecs which are conceptually the same type, but have very different implementations.

data GCodec i b d e where

Constructors

Plain :: GDecoder i d -> GEncoder b e -> GCodec i b d e

Wrap :: (d' -> Maybe d) -> (e -> Maybe e') -> GCodec i b d' e' -> GCodec i b d e

Decoders

type Decoder a = GDecoder ByteString a

A decoder is a function which, given some input bytestring, will attempt to decode it according to the construction of the decoder and return any remaining input.

type GDecoder i a = StateT i Maybe a

toDecoder :: (i -> Maybe (a, i)) -> GDecoder i a

Construct a decoder.

runDecoder :: GDecoder i a -> i -> Maybe (a, i)

Run a decoder.

decode :: i -> GCodec i b d e -> Maybe d

Run a codec's decoder.

decoderOf :: GCodec i b d e -> GDecoder i d

Get the decoder of a codec.

Encoders

type Encoder a = GEncoder Builder a

An encoder is a function which, given some input value, will attempt to encode it according to the construction of the encoder, producing a ByteString builder.

type GEncoder b a = Op (Both b) a

toEncoder :: (a -> Maybe b) -> GEncoder b a

Construct an encoder.

runEncoder :: GEncoder b a -> a -> Maybe b

Run a decoder.

encode :: e -> GCodec i b d e -> Maybe b

Run a codec's encoder.

encoderOf :: GCodec i b d e -> GEncoder b e

Get the encoder of a codec.

Both Monoid

newtype Both a

The Both Monoid is like Maybe, but requires both of its arguments to be Just or the result will be Nothing.

Constructors

Both

Fields

getBoth :: Maybe a